fixed the visualization of themes and finished the development of the menu
